+++ src/api/errors.ts
@@ -1,8 +1,11 @@
-export type SiteverifyErrorCode =
+export type APIErrorCode =
   | typeof SITEKEY_INVALID
   | typeof AUTH_INVALID
   | typeof AUTH_REQUIRED
   | typeof BAD_REQUEST
+  | typeof TOKEN_INVALID
+  | typeof TOKEN_EXPIRED
+  | typeof TOKEN_MISSING
   | typeof RESPONSE_TIMEOUT
   | typeof RESPONSE_DUPLICATE
   | typeof RESPONSE_INVALID
@@ -34,6 +37,24 @@ export const SITEKEY_INVALID = "sitekey_invalid";
  */
 export const BAD_REQUEST = "bad_request";
 /**
+ * The token you provided was invalid.
+ *
+ * HTTP status 200.
+ */
+export const TOKEN_INVALID = "token_invalid";
+/**
+ * The token has expired.
+ *
+ * HTTP status 200.
+ */
+export const TOKEN_EXPIRED = "token_expired";
+/**
+ * You forgot to add the token parameter.
+ *
+ * HTTP status 400.
+ */
+export const TOKEN_MISSING = "token_missing";
+/**
  * The response has expired.
  *
  * HTTP status 200.
@@ -64,11 +85,14 @@ export const RESPONSE_MISSING = "response_missing";
  */
 export const INTERNAL_SERVER_ERROR = "internal_server_error";
 
-export const ERROR_CODE_TO_STATUS: Record<SiteverifyErrorCode, number> = {
+export const ERROR_CODE_TO_STATUS: Record<APIErrorCode, number> = {
   [SITEKEY_INVALID]: 400,
   [AUTH_INVALID]: 401,
   [AUTH_REQUIRED]: 401,
   [BAD_REQUEST]: 400,
+  [TOKEN_INVALID]: 200,
+  [TOKEN_EXPIRED]: 200,
+  [TOKEN_MISSING]: 400,
   [RESPONSE_TIMEOUT]: 200,
   [RESPONSE_DUPLICATE]: 200,
   [RESPONSE_INVALID]: 200,