+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.0
+
+* Adds support for the Risk Intelligence Retrieve API.
+* Update dev dependencies.
+* Internal rename of `SiteverifyErrorCode` to `APIErrorCode`.
+* Rename of `VerifyClientErrorCode` to `ClientErrorCode`.
+
 ## 0.2.0
 
 * Add support for *Risk Intelligence* data in siteverify responses.